+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - navigate the current input with left/right arrow keys as well as CTRL+a/e
 - implement sending and receiving `beep` messages
+- allow changing nickname with `/name` or `/nick`
 
 # 0.2.1