+++ icb/CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.x.y (not yet released)
 
 - implement sending `beep` commands and receiving `beep` messages
+- handle nickname changes through the `name` command